Removing eval dependency by using react
