+++ package.json
@@ -15,14 +15,20 @@
   "dependencies": {
     "bulma": "^0.7.1",
     "jquery": "^3.3.1",
+    "react": "16.14.0",
+    "react-dom": "16.14.0",
     "vue": "^2.5.16"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
-    "babel-loader": "^7",
+    "babel-loader": "7.x",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-stage-2": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "del": "^3.0.0",
     "file-loader": "^0.11.2",
     "fs-extra": "^0.30.0",
+    "gulp": "^3.9.1",
     "gulp-babel": "^7.0.1",
     "gulp-imagemin": "^4.1.0",
     "gulp-inject-string": "^1.1.1",
@@ -30,9 +36,8 @@
     "gulp-rename": "^1.4.0",
     "gulp-sequence": "^1.0.0",
     "gulp-zip": "^4.1.0",
-    "gulp": "^3.9.1",
-    "webpack-cli": "3",
-    "webpack": "4"
+    "webpack": "4",
+    "webpack-cli": "3"
   },
   "engines": {
     "node": "8.x"
