+++ src/manifest-chrome.json
@@ -22,6 +22,5 @@
   "icons": {
     "128": "img/favicon-128x128.png"
   },
-  "permissions": ["storage", "tabs", "<all_urls>"],
-  "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'"
+  "permissions": ["storage", "tabs", "<all_urls>"]
 }
