+++ src/popup.html
@@ -1,13 +1,18 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <meta charset="UTF-8">
+    <meta charset="UTF-8" />
     <title>Newsit Settings</title>
-    <link rel="icon" type="image/png" sizes="32x32" href="./img/favicon-32x32.png">
-    <link rel="stylesheet" href="./vendor/bulma.min.css">
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="32x32"
+      href="./img/favicon-32x32.png"
+    />
+    <link rel="stylesheet" href="./vendor/bulma.min.css" />
     <style>
       .hero-body {
-        padding: 15px !important;  
+        padding: 15px !important;
       }
       #newsit-form {
         padding-top: 15px;
@@ -21,71 +26,12 @@
   </head>
   <body>
     <div class="columns" id="app">
-      <div class="column">
-        <section class="hero is-dark is-bold">
-          <div id="newsit-header" class="hero-body">
-            <div class="container">
-              <a href="https://newsit.benwinding.com" target="_blank">
-                <img src="img/icon.svg" style="width: 90px;">
-              </a>
-              <p class="is-pulled-right">
-              version: {{version}}
-              </p>
-            </div>
-          </div>
-        </section>
-        <section id="newsit-form" class="section">
-          <section class="form">
-            <div class="field">
-              <div class="control has-text-centered">
-                <button class="button is-info" v-bind:class="{'is-loading': isLoadingBtn}" v-on:click="onClickCheckNow">Check Current Page</button>
-              </div>
-            </div>
-            <div class="field">
-              <div class="control">
-                <label class="checkbox">
-                  <input type="checkbox" v-model="formCurrentSiteEnabled">
-                  Enabled on this website
-                </label>
-              </div>
-            </div>
-            <div class="field">
-              <div class="control">
-                <label class="checkbox">
-                  <input type="checkbox" v-model="formBtnsEnabled">
-                  Run automatically on page load
-                </label>
-              </div>
-            </div>
-            <div class="field">
-              <label class="label">Links Location</label>
-              <div class="control">
-                <div class="select">
-                  <select v-model="formBtnsLocation">
-                    <option disabled value="">Nothing selected</option>
-                    <option v-for="option in options.formBtnsLocation" v-bind:value="option.value">
-                      {{ option.text }}
-                    </option>
-                  </select>
-                </div>
-              </div>
-            </div>
-            <div class="field">
-              <label class="label">Link Size</label>
-              <div class="control">
-                <input type="range" v-model="formBtnsSize"
-                min="0.2"
-                max="2"
-                step="0.05"
-                >
-              </div>
-              {{formBtnsSize}}
-            </div>
-          </section>
-        </section>
-      </div>
     </div>
-    <script type="text/javascript" src="vendor/vue.min.js"></script>
-    <script type="text/javascript" src="js/popup.js"></script>
+    <script type="text/javascript" src="/node_modules/react/umd/react.development.js"></script>
+    <script
+      type="text/javascript"
+      src="/node_modules/react-dom/umd/react-dom.development.js"
+    ></script>
+    <script type="text/javascript" src="/build/popup.js"></script>
   </body>
 </html>
